Replace IngestTaskTest with NewIngestTaskTest
